Moved all game API calls under /game
