+++ docs/api-reference/authentication.mdx
@@ -13,7 +13,7 @@ This method provides a seamless way for new users to get started while still ens
 ### Endpoint
 
 ```http
-POST https://cratebytes.com/api/platform/guest
+POST https://cratebytes.com/api/game/platform/guest
 ```
 
 <ParamField body="projectKey (body)" type="string" required>
@@ -60,7 +60,7 @@ Steam Authentication allows users to log in to the CrateBytes API using their St
 ### Endpoint
 
 ```http
-POST https://cratebytes.com/api/platform/steam
+POST https://cratebytes.com/api/game/platform/steam
 ```
 
 <ParamField body="projectKey (body)" type="string" required>
