+++ docs/api-reference/game/metadata.mdx
@@ -10,7 +10,7 @@ description: "Store and retrieve player metadata with the CrateBytes API."
 ### Endpoint
 
 ```http
-GET https://cratebytes.com/api/game/metadata/get
+GET https://cratebytes.com/api/game/game/metadata/get
 ```
 
 <ParamField header="authorization (header)" type="string" required>
@@ -31,7 +31,7 @@ GET https://cratebytes.com/api/game/metadata/get
 ### Endpoint
 
 ```http
-PUT https://cratebytes.com/api/game/metadata/add
+PUT https://cratebytes.com/api/game/game/metadata/add
 ```
 
 <ParamField header="authorization (header)" type="string" required>
@@ -56,7 +56,7 @@ PUT https://cratebytes.com/api/game/metadata/add
 ### Endpoint
 
 ```http
-DELETE https://cratebytes.com/api/game/metadata/delete
+DELETE https://cratebytes.com/api/game/game/metadata/delete
 ```
 
 <ParamField header="authorization (header)" type="string" required>
