+++ docs/api-reference/game/session.mdx
@@ -12,7 +12,7 @@ The `Start Session` endpoint is used to initiate a new game session for a player
 ### Endpoint
 
 ```http
-POST https://cratebytes.com/api/game/gameplay/start
+POST https://cratebytes.com/api/game/game/gameplay/start
 ```
 
 <ParamField header="authorization (header)" type="string" required>
@@ -44,7 +44,7 @@ The `Heartbeat` endpoint is used to keep an active game session alive. This is c
 ### Endpoint
 
 ```http
-POST https://cratebytes.com/api/game/gameplay/heartbeat
+POST https://cratebytes.com/api/game/game/gameplay/heartbeat
 ```
 
 <ParamField header="authorization (header)" type="string" required>
@@ -86,7 +86,7 @@ The `End Session` endpoint is used to terminate an active game session for a pla
 ### Endpoint
 
 ```http
-POST https://cratebytes.com/api/game/gameplay/end
+POST https://cratebytes.com/api/game/game/gameplay/end
 ```
 
 <ParamField header="authorization (header)" type="string" required>
